[fix] GNB 파일 이동 및 수정
