+++ src/RootLayout.tsx
@@ -1,12 +1,12 @@
-import { Outlet } from "react-router-dom"
+import { Outlet } from 'react-router-dom';
 
-import GlobalNavigationBar from "./components/GlobalNavigationBar"
+import GlobalNavigationBar from './components/GlobalNavigationBar';
 
 export default function RootLayout() {
   return (
-    <div className='w-full h-full relative'>
+    <div className="w-full h-full relative">
       <Outlet />
       <GlobalNavigationBar />
     </div>
-  )
+  );
 }
