+++ src/components/GlobalNavigationBar.tsx
@@ -3,7 +3,7 @@ import { Dispatch, useState } from 'react';
 import { Link } from 'react-router-dom';
 
 type RouteItem = {
-  name: string;
+  text: string;
   route: string;
   icon: string;
   iconFill: string;
@@ -11,32 +11,32 @@ type RouteItem = {
 
 const ROUTER_STATE: RouteItem[] = [
   {
-    name: '홈',
+    text: '홈',
     route: '/',
     icon: 'bg-home-icon',
     iconFill: 'bg-home-fill-icon',
   },
   {
-    name: '검색하기',
-    route: 'search',
+    text: '검색하기',
+    route: '/search',
     icon: 'bg-search-icon',
     iconFill: 'bg-search-fill-icon',
   },
   {
-    name: '생성하기',
-    route: 'create',
+    text: '생성하기',
+    route: '/create',
     icon: 'bg-add-icon',
     iconFill: 'bg-create-fill-icon',
   },
   {
-    name: '북마크',
-    route: 'bookmark',
+    text: '북마크',
+    route: '/bookmark',
     icon: 'bg-bookmark-icon',
     iconFill: 'bg-bookmark-fill-icon',
   },
   {
-    name: '마이페이지',
-    route: 'user',
+    text: '마이페이지',
+    route: '/user',
     icon: 'bg-person-icon',
     iconFill: 'bg-person-fill-icon',
   },
@@ -46,7 +46,7 @@ type GNBButtonProps = {
   route: string;
   icon: string;
   iconFill: string;
-  name: string;
+  text: string;
   currentPage: string;
   setPage: Dispatch<SetStateAction<string>>;
 };
@@ -55,29 +55,37 @@ function GNBButton({
   route,
   icon,
   iconFill,
-  name,
+  text,
   currentPage,
-  setPage
+  setPage,
 }: GNBButtonProps) {
+  const renderIcon = () => {
+    if (currentPage === route) return iconFill;
+    return icon;
+  };
+
   return (
-    <li className="flex basis-full bg-red-100">
+    <li className="flex basis-full">
       <Link
         to={route}
         onClick={() => {
-          setPage(route)
+          setPage(route);
         }}
-        className={`h-full w-full bg-no-repeat bg-[center_top_10px] ${currentPage === route ? iconFill : icon}`}
+        className={`h-full w-full bg-no-repeat bg-[center_top_10px] ${renderIcon()}`}
       >
-        <p className="sr-only">{name}</p>
+        <p className="sr-only">{text}</p>
       </Link>
     </li>
   );
 }
 
 export default function GlobalNavigationBar() {
-  const [currentPage, setCurrentPage] = useState<string>('홈');
+  const [currentPage, setCurrentPage] = useState<string>(
+    window.location.pathname
+  );
+
   return (
-    <nav className="absolute bottom-0 w-full h-80pxr px-14pxr pb-24pxr">
+    <nav className="absolute bottom-0 w-full h-80pxr px-side pb-24pxr">
       <ul className="flex flex-row list-none w-full h-full">
         {ROUTER_STATE.map((item, idx) => {
           return (
@@ -93,4 +101,3 @@ export default function GlobalNavigationBar() {
     </nav>
   );
 }
-
